:zap: clean up
